refactor: split rollsite into multiple files according to their functional
